:recycle: Rename Contact model
