+++ lib/core/models/contact.dart
